LayoutTests/platform/gtk/fast/events/mouse-relative-position-expected.txt

CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
CONSOLE MESSAGE: event.layerX and event.layerY are broken and deprecated in WebKit. They will be removed from the engine in the near future.
span

PASS simulateElementClick(testElement, [100, 100, 200, 200]); is '100, 100, 200, 200'
FAIL simulateElementClick(spanElement, [16, 10, 16, 10]); should be 16, 10, 16, 10. Was 12, 10, 112, 110.
FAIL simulateElementClick(inputElement, [40, 10, 40, 10]); should be 40, 10, 40, 10. Was 40, 10, 95, 10.
